+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ai-sdk/angular
 
+## 2.0.104
+
+### Patch Changes
+
+- ai@6.0.103
+
+## 2.0.103
+
+### Patch Changes
+
+- ai@6.0.102
+
+## 2.0.102
+
+### Patch Changes
+
+- Updated dependencies [5230482]
+  - ai@6.0.101
+
+## 2.0.101
+
+### Patch Changes
+
+- Updated dependencies [b7fba77]
+  - ai@6.0.100
+
 ## 2.0.100
 
 ### Patch Changes