+++ packages/react/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ai-sdk/react
 
+## 3.0.105
+
+### Patch Changes
+
+- ai@6.0.103
+
+## 3.0.104
+
+### Patch Changes
+
+- ai@6.0.102
+
+## 3.0.103
+
+### Patch Changes
+
+- Updated dependencies [5230482]
+  - ai@6.0.101
+
+## 3.0.102
+
+### Patch Changes
+
+- Updated dependencies [b7fba77]
+  - ai@6.0.100
+
 ## 3.0.101
 
 ### Patch Changes