+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,32 @@
 
 ### Patch Changes
 
+- ai@6.0.103
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.102
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5230482]
+  - ai@6.0.101
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [b7fba77]
+  - ai@6.0.100
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.99
 
 ## 0.0.1